*영구삭제시(파일) 파일인지 체크

+++ src/main/java/egovframework/ezEKP/ezWebFolder/service/impl/EzWebFolderServiceimpl_m.java
@@ -887,7 +887,7 @@ public class EzWebFolderServiceimpl_m implements EzWebFolderService_m {
 		
 		LOGGER.debug("pDirPath=" + pDirPath);
 		
-		if (file.exists()) {
+		if (file.exists() && file.isFile()) {
 			if (file.delete()) {
 				isDeleted = 1;
 				ezWebFolderService.saveLog("P", userInfo.getCompanyID(), userInfo.getOffset(), userInfo.getId(), userName1, userName2,
